Eliminates circular dependency between ConfigCatClientCache and ConfigCatClient modules
