+++ src/ConfigCatClient.ts
@@ -8,7 +8,6 @@ import { User, IRolloutEvaluator, RolloutEvaluator } from "./RolloutEvaluator";
 import { Setting, RolloutRule, RolloutPercentageItem, ConfigFile } from "./ProjectConfig";
 import { OverrideBehaviour } from "./FlagOverrides";
 import { getSettingsFromConfig } from "./Utils";
-import { ConfigCatClientCache } from "./ConfigCatClientCache";
 import { isWeakRefAvailable } from "./Polyfills";
 
 export interface IConfigCatClient {
@@ -64,6 +63,55 @@ export interface IConfigCatClient {
 
     /** Clears the default user. */
     clearDefaultUser(): void;
+}
+
+export class ConfigCatClientCache {
+    private instances: Record<string, [WeakRef<ConfigCatClient>, object]> = {};
+
+    public getOrCreate(options: ConfigCatClientOptions, configCatKernel: IConfigCatKernel): [ConfigCatClient, boolean] {
+        let instance: ConfigCatClient | undefined;
+
+        const cachedInstance = this.instances[options.apiKey];
+        if (cachedInstance) {
+            const [weakRef] = cachedInstance;
+            instance = weakRef.deref();
+            if (instance) {
+                return [instance, true];
+            }
+        }
+
+        const token = {};
+        instance = new ConfigCatClient(options, configCatKernel, token);
+        this.instances[options.apiKey] = [new WeakRef(instance), token];
+        return [instance, false];
+    }
+
+    public remove(sdkKey: string, cacheToken: object): boolean {
+        const cachedInstance = this.instances[sdkKey];
+
+        if (cachedInstance) {
+            const [weakRef, token] = cachedInstance;
+            const instanceIsAvailable = !!weakRef.deref();
+            if (!instanceIsAvailable || token === cacheToken) {
+                delete this.instances[sdkKey];
+                return instanceIsAvailable;
+            }
+        }
+
+        return false;
+    }
+
+    public clear(): ConfigCatClient[] {
+        const removedInstances: ConfigCatClient[] = [];
+        for (let [sdkKey, [weakRef]] of Object.entries(this.instances)) {
+            let instance = weakRef.deref();
+            if (instance) {
+                removedInstances.push(instance);
+            }
+            delete this.instances[sdkKey];
+        }
+        return removedInstances;
+    }
 }
 
 const clientInstanceCache = new ConfigCatClientCache();
